+++ src/objects/sphere.h
@@ -18,12 +18,12 @@
 
 typedef struct	s_sphere {
 	t_vec3f		center;
-	float		radius;
-	float		radius2;
+	double		radius;
+	double		radius2;
 }				t_sphere;
 
-t_sphere		*sphere_new(t_vec3f center, float radius);
-int				sphere_intersect(t_sphere *sphere, t_ray *hit, float *res);
+t_sphere		*sphere_new(t_vec3f center, double radius);
+int				sphere_intersect(t_sphere *sphere, t_ray *hit, double *res);
 void			sphere_print(t_sphere *sphere);
 void			sphere_normale(t_sphere *sphere, t_ray *hit);
